product and brand update ok

+++ src/components/Brand/AllBrand/AllBrand.jsx
@@ -1,22 +1,40 @@
 import { useState } from "react";
 import { useLoaderData } from "react-router-dom";
 import AllBrandCard from "../AllBrandCard/AllBrandCard";
+import Swal from "sweetalert2";
 
 const AllBrand = () => {
   const allbrand = useLoaderData();
   const [brands, setBrand] = useState(allbrand)
 
     const handleDeleteBrand=(_id)=>{
-      console.log('delete btn click', _id);
-      fetch(`https://brand-shop-server-one-bice.vercel.app/${_id}`, {
-        method: 'DELETE', 
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          
+          fetch(`https://brand-shop-server-one-bice.vercel.app/brand/${_id}`, {
+            method: 'DELETE', 
+          })
+          .then(res=>res.json())
+          .then(()=>{
+            Swal.fire(
+              'Deleted!',
+              'Your brand has been deleted.',
+              'success'
+            )
+            const remainingBrand = brands.filter(brand=>brand._id !== _id)
+            setBrand(remainingBrand)
+          })
+        }
       })
-      .then(res=>res.json())
-      .then(()=>{
-    
-        const remainingBrand = brands.filter(brand=>brand._id !== _id)
-        setBrand(remainingBrand)
-      })
+     
     }
   return (
     <div>
